firmware: version "anode"
average voltage is now constantly computed
as long as there is nothing available on serial.

this allows decoupling voltage computing & serial
communication, which is now almost instand

I don't know that the math is good, it probably isnt.
I'm using some kind of moving average to reduce
computation: every chunk of 50 voltage readings,
you add the average of this to another chunk of 50 averages,
and you average on that.

3 params relevant to this mechanics:
  - the base chunk size,
  - the chunk of chunks size,
  - the sleep delay in main loop